requirements.txt
python-telegram-bot==21.6
google-genai>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
httpx>=0.26.0
payos>=1.1.0
